% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avallecam.R
\docType{package}
\name{avallecam}
\alias{avallecam}
\alias{avallecam-package}
\title{avallecam package}
\description{
avallecam - custom functions to solve current problems
}
\author{
Andree Valle Campos \email{avallecam@gmail.com}
}
