% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tidymodel_pr.R
\name{epi_tidymodel_pr}
\alias{epi_tidymodel_pr}
\title{tidy up -even more- an output model from broom}
\usage{
epi_tidymodel_pr(wm1)
}
\arguments{
\item{wm1}{model output from glm}
}
\value{
dataset with exponentiated estimate and confidence intervals.
}
\description{
Prepares the output to a printable table format. Exclusive for models that estimate prevalence ratios (PR)
}
