% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tidymodel_rr.R
\name{epi_tidymodel_rr}
\alias{epi_tidymodel_rr}
\title{tidy up -even more- an output model from broom}
\usage{
epi_tidymodel_rr(wm1, digits = 5)
}
\arguments{
\item{wm1}{model output from glm}

\item{digits}{number of decimal digits for pvalue}
}
\value{
dataset with exponentiated estimate and confidence intervals.
}
\description{
Prepares the output to a printable table format. Exclusive for models that estimate risk ratios (RR)
}
