% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avmisc.R
\name{print_inf}
\alias{print_inf}
\alias{read_lastfile}
\title{The Most Important Custom Functions}
\usage{
print_inf(object)

read_lastfile(path, pattern)
}
\arguments{
\item{object}{objeto input}

\item{path}{path to folder}

\item{pattern}{any specific file extension}
}
\value{
Really (context dependent) important outputs.
}
\description{
benchwork optimizer
}
\section{Functions}{
\itemize{
\item \code{print_inf}: make a quick(er) \code{print(n=Inf)}

\item \code{read_lastfile}: read the last file in a folder (ideal for workflows with daily inputs and updates)
}}

\examples{

\dontrun{

# print_inf ---------------------------------------------------

library(tidyverse)
mtcars \%>\% as_tibble() \%>\% print_inf()

# read_lastfile------------------------------------------------

# # read last name in a long list of files
# lastfile_name <- read_lastfile(path = "../../long_rute_to_folder/", pattern = "*.xlsx") \%>\% print()
#
# # read that filename
# read_excel(lastfile_name)

}

}
