% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_gpx}
\alias{read_gpx}
\title{Read gpx files}
\usage{
read_gpx(list)
}
\arguments{
\item{list}{list of gpx files. easier if they came from a list.files() output}
}
\value{
dataset of geolocation points with waypoint name and complementary information.
}
\description{
Read gpx files from different GPS tracking devices
}
