% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatially_useful.R
\name{st_coordinates_tidy}
\alias{st_coordinates_tidy}
\title{retrieve coordinates within the original sf/data.frame object}
\usage{
st_coordinates_tidy(sf_object)
}
\arguments{
\item{sf_object}{sf_object input}
}
\value{
sf and data.frame object X and Y extracted from geometry column
}
\description{
retrieve coordinates within the original sf/data.frame object. an alternative from st_coordinates() (https://r-spatial.github.io/sf/reference/st_coordinates.html)
}
\section{Functions}{
\itemize{
\item \code{st_coordinates_tidy}: retrieve coordinates within the original sf/data.frame object. an alternative from st_coordinates() (https://r-spatial.github.io/sf/reference/st_coordinates.html)
}}

\examples{

library(tidyverse)
library(sf)
library(avallecam)

sites <- tibble(gpx_point = c("a","b"),
                longitude = c(-80.144005, -80.109),
                latitude = c(26.479005,26.83)) \%>\% print()

sites_sf <- sites \%>\%
  st_as_sf(coords = c("longitude", "latitude"),
           remove = T,
           crs = 4326) \%>\% print()

sites_sf \%>\%
  st_coordinates()

sites_sf \%>\%
  st_coordinates_tidy()

}
