% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabyl_adorn.R
\name{adorn_2x2}
\alias{adorn_2x2}
\title{adorn up a tabyl output}
\usage{
adorn_2x2(tabyl)
}
\arguments{
\item{tabyl}{tabyl input}
}
\value{
tabyl object with horizontal % and in front N.
}
\description{
Creates the usual 2 by 2 output to a tabyl format.
}
