% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tidymodel.R
\name{epi_tidymodel_or}
\alias{epi_tidymodel_or}
\alias{epi_tidymodel_rr}
\alias{epi_tidymodel_pr}
\alias{epi_tidymodel_up}
\alias{epi_tidynested}
\alias{epi_tidymodel_coef}
\title{Tidy up -even more- an output model using broom functions}
\usage{
epi_tidymodel_or(model_output, digits = 5)

epi_tidymodel_rr(model_output, digits = 5)

epi_tidymodel_pr(model_output, digits = 5)

epi_tidymodel_up(reference_model, variable)

epi_tidynested(add_nested, level = i)

epi_tidymodel_coef(model_output, digits = 5)
}
\arguments{
\item{model_output}{output raw GLM model}

\item{digits}{digits for numeric double objects (p.value)}

\item{reference_model}{reference model to update}

\item{variable}{new variable to update into the model}

\item{add_nested}{add1() output model}

\item{level}{level of nesting process}
}
\value{
table summary from traditional epi models.
}
\description{
Summarize OR, RR, PR regression outputs and update models for simple and other multiple models.
}
\section{Functions}{
\itemize{
\item \code{epi_tidymodel_or}: summarize and calculates OR coefficients from bin-logit regression

\item \code{epi_tidymodel_rr}: summarize and calculates RR coefficients from bin-log regression

\item \code{epi_tidymodel_pr}: summarize and calculates PR coefficients from bin/poisson-log regression

\item \code{epi_tidymodel_up}: updates reference models (null or parsimonius) by adjusting with a new covariate. to use with purrr::map() and requires to use rlang::sym() to recognize the variable.

\item \code{epi_tidynested}: tidy output for each update level in the nested models procedure

\item \code{epi_tidymodel_coef}: summarize and calculates coefficients from linear regression (gaussian identity GLM)
}}

\examples{

# ejemplo
# falta

}
