% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_tidy.R
\name{get_distance_m}
\alias{get_distance_m}
\title{distance tibble}
\usage{
get_distance_m(set1_lat, set1_lon, set2_lat, set2_lon)
}
\arguments{
\item{set1_lat}{set1 of latitude coordinates}

\item{set1_lon}{set1 of longitude coordinates}

\item{set2_lat}{set2 of latitude coordinates}

\item{set2_lon}{set2 of longitude coordinates}
}
\value{
distance as double vector
}
\description{
distance tibble
}
\section{Functions}{
\itemize{
\item \code{get_distance_m}: generates a distance output between two set of points within a tibble and flexible with group_by and purrr::pmap
}}

\examples{

library(tidyverse)
library(geosphere)
library(purrr)
library(avallecam)

d <- tibble(run = c(rep(1,6), rep(2,6)),
            grp = c(rep("a",3),rep("b",3),rep("a",3),rep("b",3)),
            lat = rep(c(57.15508, 52.41283, 57.15521,
                        52.41278, 57.14520, 52.41317),2),
            lon = rep(c(-2.07887, -4.07806, -2.07886,
                        -4.07803, -2.07886, -4.07858),2)) \%>\% print()
#test
d \%>\%
  group_by(run,grp) \%>\%
  mutate(dist=pmap_dbl(list(lat,lon,lag(lat),lag(lon)),get_distance_m))

}
